backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
pandas
python-dotenv
google-generativeai
speechrecognition
pydub
python-multipart
